msm8953-common: ipa: check return status of HAL call to avoid transaction failures
If the underlying HAL service object is invalidated and we did not get
a chance to invalidate it internally we may hit HAL transaction failure
error and the process would be forced to abort. To prevent the
transaction error we need to use the isOk check right after the HAL API
is invoked to check whether the HAL transaction was successful.

Change-Id: Idd21209f5c0407b75c2637190df5ba8158ef1a80
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>